backend/pyproject.toml
[project]
name = "backend"
version = "0.1.0"
description = ""
authors = [
    {name = "mattia federici",email = "mattia.federici@invariant.ai"}
]
readme = "README.md"
requires-python = ">=3.10,<4.0"
dependencies = [
    "fastapi (>=0.117.1,<0.118.0)",
    "uvicorn (>=0.37.0,<0.38.0)",
    "pyotp (>=2.9.0,<3.0.0)",
    "qrcode[pil] (>=8.2,<9.0)",
    "python-multipart (>=0.0.20,<0.0.21)",
    "loguru (>=0.7.3,<0.8.0)"
    ,"PyJWT (>=2.8.0,<3.0.0)",
    "python-jose (>=3.3.0,<4.0.0)"
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
